+++ zio-http/src/test/scala/zio/http/model/SchemeSpec.scala
@@ -18,22 +18,5 @@ object SchemeSpec extends ZIOSpecDefault {
     test("null string decode") {
       assert(Scheme.decode(null))(isNone)
     },
-    test("java http scheme") {
-      checkAll(jHttpScheme) { jHttpScheme =>
-        assertTrue(Scheme.fromJScheme(jHttpScheme).flatMap(_.toJHttpScheme).get == jHttpScheme)
-      }
-    },
-    test("java websocket scheme") {
-      checkAll(jWebSocketScheme) { jWebSocketScheme =>
-        assertTrue(
-          Scheme.fromJScheme(jWebSocketScheme).flatMap(_.toJWebSocketScheme).get == jWebSocketScheme,
-        )
-      }
-    },
   )
-
-  private def jHttpScheme: Gen[Any, HttpScheme] = Gen.fromIterable(List(HttpScheme.HTTP, HttpScheme.HTTPS))
-
-  private def jWebSocketScheme: Gen[Any, WebSocketScheme] =
-    Gen.fromIterable(List(WebSocketScheme.WS, WebSocketScheme.WSS))
 }
